add saving memory in JSON
