+++ app/build.gradle.kts
@@ -50,4 +50,5 @@ dependencies {
     implementation("com.google.android.material:material:1.12.0")
     implementation("androidx.fragment:fragment-ktx:1.8.1")
     implementation("androidx.activity:activity-ktx:1.9.0")
+    implementation("com.google.code.gson:gson:2.11.0")
 }
